+++ packages/fuels-test-helpers/Cargo.toml
@@ -13,7 +13,7 @@ fuel-core = { version = "0.8", default-features = false, optional = true}
 
 fuel-gql-client = { version = "0.8", default-features = false }
 fuels-core = { version = "0.14.1", path = "../fuels-core" }
-fuels-signers = { version = "0.14.1", path = "../fuels-signers"}
+fuels-signers = { version = "0.14.1", path = "../fuels-signers", optional = true }
 rand = { version = "0.8.4", default-features = false }
 tokio = "1.15"
 
@@ -28,6 +28,6 @@ tempfile = "3.0.1"
 
 
 [features]
-#default = ["fuels-signers"]
+default = ["fuels-signers"]
 fuel-core-bin = ["fuel-core-interfaces"]
 fuel-core-lib = ["fuel-core"]
