+++ packages/fuels-test-helpers/src/signers.rs
@@ -11,10 +11,8 @@ use fuel_core_interfaces::model::Coin;
 use fuels_signers::{provider::Provider, LocalWallet, Signer};
 
 use crate::{
-    setup_single_asset_coins,
-    setup_test_client,
-    wallets_config::WalletsConfig,
-    // DEFAULT_COIN_AMOUNT, DEFAULT_NUM_COINS,
+    setup_single_asset_coins, setup_test_client, wallets_config::WalletsConfig,
+    DEFAULT_COIN_AMOUNT, DEFAULT_NUM_COINS,
 };
 
 #[cfg(feature = "fuel-core-lib")]
@@ -82,13 +80,16 @@ pub async fn setup_test_provider(
 
 #[cfg(not(feature = "fuel-core-lib"))]
 pub async fn launch_provider_and_get_single_wallet() -> LocalWallet {
-    let mut wallets = launch_provider_and_get_wallets(WalletsConfig::new_single(None, None)).await;
+    let mut wallets =
+        launch_provider_and_get_wallets(WalletsConfig::new_single(None, None)).await;
 
     wallets.pop().unwrap()
 }
 
 #[cfg(not(feature = "fuel-core-lib"))]
-pub async fn launch_provider_and_get_wallets(config: WalletsConfig) -> Vec<LocalWallet> {
+pub async fn launch_provider_and_get_wallets(
+    config: WalletsConfig,
+) -> Vec<LocalWallet> {
     let mut wallets: Vec<LocalWallet> = (1..=config.num_wallets)
         .map(|_i| LocalWallet::new_random(None))
         .collect();
@@ -114,7 +115,9 @@ pub async fn launch_provider_and_get_wallets(config: WalletsConfig) -> Vec<Local
 }
 
 #[cfg(not(feature = "fuel-core-lib"))]
-pub async fn setup_test_provider(coins: Vec<(UtxoId, Coin)>) -> (Provider, SocketAddr) {
+pub async fn setup_test_provider(
+    coins: Vec<(UtxoId, Coin)>,
+) -> (Provider, SocketAddr) {
     let (client, addr) = setup_test_client(coins).await;
     (Provider::new(client), addr)
 }
